etl/requirements.txt
asgiref~=3.6.0
attrs==23.2.0
click==8.1.7
elasticsearch==8.10.0
eradicate==2.3.0
mccabe==0.7.0
mr-proper==0.0.7
psycopg2-binary==2.9.9
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==1.0.1
sqlparse==0.4.4
stdlib-list==0.10.0
typing_extensions==4.9.0
pydantic~=1.10.9
pip~=23.2
zipp~=1.0.0
toml~=0.10.2
wheel~=0.37.1
cryptography~=3.4.8
py~=1.10.0
lxml~=4.8.0
Pillow~=9.0.1
keyring~=23.5.0
colorama~=0.4.4
setuptools~=59.6.0
PyYAML~=5.3.1
numpy~=1.24.2
requests~=2.28.2
future~=0.18.2